This is the script
This is the first version. There is still much to correct. The script is not general enough to be implemented easily. Some changes are required. Also, some of the methodes used are clumsy. For example, when constructing an array for February based on wether this year is a leap year or not, why must the script change the whole array? It's wastefull! Why not simply push one more cell to the array?

Other issues:
- Some global variables are not defined (although you only need to remove // comments to define them)
- Why do I waste 31 lines of code for January array? Why not construc it with loop?
- path for the image of Finnish flag is defined but the relative paht might not be same for everybody
- variable eil defines the text which is displayd when there is not flag flying day. There is a class on <a href> element and <span> element which are required for css tooltip. But what if you don't have that in your css? Should I comment instructions for css or would it be better to remove the